lab3/src/main/java/ie/spring/lab1/LanguageConfig.java
package ie.spring.lab1;

import org.springframework.context.MessageSource;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.support.ResourceBundleMessageSource;

@Configuration
public class LanguageConfig {

  @Bean
  public MessageSource messageSource() {
    ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
    messageSource.setBasename("messages"); // Base name for your language files
    messageSource.setDefaultEncoding("UTF-8");
    return messageSource;
  }
}
